requirements.txt
dash==2.0.0
dash-bootstrap-components==1.0.0
python-binance==1.0.15
